+++ modules/configure_client.py
@@ -15,7 +15,7 @@ import paramiko
 
 from . import _utils
 
-CLIENT_UNZIPPED_DIR = 'PUSH_Clients-master'
+CLIENT_UNZIPPED_DIR = "PUSH_Clients-master"
 
 
 def _add_metric(archive_dir, client_path, name, client):
@@ -26,60 +26,58 @@ def _add_metric(archive_dir, client_path, name, client):
     client. Adds the metric that is enabled by the user in manage_checks.py
     """
 
-    if '@' in client_path:
-        client_path = client_path.split(':')[-1]
+    if "@" in client_path:
+        client_path = client_path.split(":")[-1]
 
-    if client == 'POSIX':
+    if client == "POSIX":
         filename = join(archive_dir, "POSIX/NodePingPUSHClient/moduleconfig")
         metric_path = join(
-            client_path, 'NodePingPUSHClient/modules/{0}/{0}.sh\n'.format(name))
+            client_path, "NodePingPUSHClient/modules/{0}/{0}.sh\n".format(name)
+        )
 
         string = "{0}={1}".format(name, metric_path)
 
         string = _utils.win_to_other_path(string)
 
         # Append the location of the module to the moduleconfig file
-        with open(filename, 'a', newline='\n') as f:
+        with open(filename, "a", newline="\n") as f:
             f.write(string)
 
         # Make the module executable
         sh_script = join(
-            archive_dir, "POSIX/NodePingPUSHClient/modules/{0}/{0}.sh".format(name))
+            archive_dir, "POSIX/NodePingPUSHClient/modules/{0}/{0}.sh".format(name)
+        )
 
         st = os.stat(sh_script)
         os.chmod(sh_script, st.st_mode | stat.S_IEXEC)
 
     # Use configparser to add the metric to the [modules] section
-    elif client == 'Python' or client == 'Python3':
-        filename = join(
-            archive_dir, "{0}/NodePing{0}PUSH/config.ini".format(client))
+    elif client == "Python" or client == "Python3":
+        filename = join(archive_dir, "{0}/NodePing{0}PUSH/config.ini".format(client))
 
         config = configparser.ConfigParser(allow_no_value=True)
         config.read(filename)
 
-        config.set('modules', name, value='yes')
+        config.set("modules", name, value="yes")
         _utils.write_config(config, filename)
-    elif client == 'PowerShell':
+    elif client == "PowerShell":
         filename = join(
-            archive_dir, "PowerShell/NodePingPowerShellPUSH/moduleconfig.json")
+            archive_dir, "PowerShell/NodePingPowerShellPUSH/moduleconfig.json"
+        )
 
         module = "modules\\{0}\\{0}.ps1".format(name)
 
         # Data for the new module that will go in the moduleconfig file for PS
-        array = {
-            "name": name,
-            "FileName": "powershell.exe",
-            "Arguments": module
-        }
+        array = {"name": name, "FileName": "powershell.exe", "Arguments": module}
 
-        with open(filename, 'r') as f:
+        with open(filename, "r") as f:
             json_data = json.load(f)
 
         # Append the new module info to the config dictionary
         json_data.append(array)
 
         # Dump the modules dictionary over the file in JSON format
-        with open(filename, 'w') as f:
+        with open(filename, "w") as f:
             f.write(json.dumps(json_data, indent=8, sort_keys=True))
 
 
@@ -90,27 +88,26 @@ def _empty_config_file(path, client):
     specified checks they want to enable
     """
 
-    if client == 'POSIX':
+    if client == "POSIX":
         filename = join(path, "POSIX/NodePingPUSHClient/moduleconfig")
 
-        open(filename, 'w').close()
+        open(filename, "w").close()
 
-    elif client == 'Python' or client == 'Python3':
+    elif client == "Python" or client == "Python3":
         filename = join(path, "{0}/NodePing{0}PUSH/config.ini".format(client))
 
         config = configparser.ConfigParser(allow_no_value=True)
         config.read(filename)
 
-        config.remove_option('modules', 'diskfree')
-        config.remove_option('modules', 'load')
-        config.remove_option('modules', 'memfree')
+        config.remove_option("modules", "diskfree")
+        config.remove_option("modules", "load")
+        config.remove_option("modules", "memfree")
 
         _utils.write_config(config, filename)
-    elif client == 'PowerShell':
-        filename = join(
-            path, "PowerShell/NodePingPowerShellPUSH/moduleconfig.json")
+    elif client == "PowerShell":
+        filename = join(path, "PowerShell/NodePingPowerShellPUSH/moduleconfig.json")
 
-        with open(filename, 'w') as f:
+        with open(filename, "w") as f:
             f.write("[]\n")
 
 
@@ -123,38 +120,38 @@ def _get_save_directory(check_id):
 
     check_questions = [
         {
-            'type': 'confirm',
-            'name': 'run_remote',
-            'message': 'Will the client run on a remote machine?'
+            "type": "confirm",
+            "name": "run_remote",
+            "message": "Will the client run on a remote machine?",
         },
         {
-            'type': 'input',
-            'name': 'remote_user',
-            'message': 'What remote user you will use for remote copy?',
-            'when': lambda answers: answers.get('run_remote', True)
+            "type": "input",
+            "name": "remote_user",
+            "message": "What remote user you will use for remote copy?",
+            "when": lambda answers: answers.get("run_remote", True),
         },
         {
-            'type': 'input',
-            'name': 'remote_host',
-            'message': 'What is the IP/FQDN of the remote server?',
-            'when': lambda answers: answers.get('run_remote', True)
+            "type": "input",
+            "name": "remote_host",
+            "message": "What is the IP/FQDN of the remote server?",
+            "when": lambda answers: answers.get("run_remote", True),
         },
         {
-            'type': 'input',
-            'name': 'save_destionation',
-            'message': 'What directory do you want the client stored in?'
-        }
+            "type": "input",
+            "name": "save_destionation",
+            "message": "What directory do you want the client stored in?",
+        },
     ]
 
     answers = _utils.confirm_choice(check_questions)
 
-    if not answers['run_remote']:
-        save_destionation = expanduser(answers['save_destionation'])
+    if not answers["run_remote"]:
+        save_destionation = expanduser(answers["save_destionation"])
         return "{0}/{1}".format(save_destionation, check_id)
     else:
-        user = answers['remote_user']
-        server = answers['remote_host']
-        save_dir = expanduser(answers['save_destionation'])
+        user = answers["remote_user"]
+        server = answers["remote_host"]
+        save_dir = expanduser(answers["save_destionation"])
         path = "{0}@{1}:{2}/{3}".format(user, server, save_dir, check_id)
 
         return path
@@ -171,16 +168,16 @@ def _place_client(src, dest, client, _id):
 
     _utils.seperator()
 
-    if client == 'POSIX':
+    if client == "POSIX":
         client_dir = join(src, "POSIX/NodePingPUSHClient")
     else:
         client_dir = join(src, "{0}/NodePing{0}PUSH".format(client))
 
     if "@" in dest:
-        username = dest.split('@')[0]
-        host = dest.split('@')[1].split(':')[0]
+        username = dest.split("@")[0]
+        host = dest.split("@")[1].split(":")[0]
         # Destination dir being copied to
-        dirname = dest.split('@')[1].split(':')[1]
+        dirname = dest.split("@")[1].split(":")[1]
 
         connected = False
 
@@ -190,49 +187,49 @@ def _place_client(src, dest, client, _id):
 
             ssh_questions = [
                 {
-                    'type': 'list',
-                    'name': 'use_password',
-                    'message': 'Will you use ssh keys or a password',
-                    'choices': ['keys', 'password']
+                    "type": "list",
+                    "name": "use_password",
+                    "message": "Will you use ssh keys or a password",
+                    "choices": ["keys", "password"],
                 },
                 {
-                    'type': 'password',
-                    'name': 'password',
-                    'message': 'Enter the remote user\'s password',
-                    'when': lambda answers: answers['use_password'] is 'password'
+                    "type": "password",
+                    "name": "password",
+                    "message": "Enter the remote user's password",
+                    "when": lambda answers: answers["use_password"] is "password",
                 },
                 {
-                    'type': 'input',
-                    'name': 'ssh_key',
-                    'message': 'Specify the path to the ssh keys to use',
-                    'when': lambda answers: answers['use_password'] is 'keys'
+                    "type": "input",
+                    "name": "ssh_key",
+                    "message": "Specify the path to the ssh keys to use",
+                    "when": lambda answers: answers["use_password"] is "keys",
                 },
                 {
-                    'type': 'password',
-                    'name': 'ssh_key_pass',
-                    'message': 'Password for ssh key (if one exists)',
-                    'when': lambda answers: answers['use_password'] is 'keys'
+                    "type": "password",
+                    "name": "ssh_key_pass",
+                    "message": "Password for ssh key (if one exists)",
+                    "when": lambda answers: answers["use_password"] is "keys",
                 },
                 {
-                    'type': 'input',
-                    'name': 'ssh_port',
-                    'message': 'Specify the ssh port (Default 22)',
-                    'default': '22'
+                    "type": "input",
+                    "name": "ssh_port",
+                    "message": "Specify the ssh port (Default 22)",
+                    "default": "22",
                 },
                 {
-                    'type': 'list',
-                    'name': 'remote_os',
-                    'message': 'What is the OS the client will be copied to',
-                    'choices': ['Windows', 'Other']
-                }
+                    "type": "list",
+                    "name": "remote_os",
+                    "message": "What is the OS the client will be copied to",
+                    "choices": ["Windows", "Other"],
+                },
             ]
 
             answers = _utils.confirm_choice(ssh_questions)
 
-            if not answers['ssh_port']:
+            if not answers["ssh_port"]:
                 port = 22
             else:
-                port = int(answers['ssh_port'])
+                port = int(answers["ssh_port"])
 
             try:
                 transport = paramiko.Transport((host, port))
@@ -244,14 +241,13 @@ def _place_client(src, dest, client, _id):
                 print(connect_answer)
 
                 if not connect_answer:
-                    return answers['remote_os']
+                    return answers["remote_os"]
 
                 continue
 
-            if answers['use_password'] == 'password':
+            if answers["use_password"] == "password":
                 try:
-                    transport.connect(username=username,
-                                      password=answers['password'])
+                    transport.connect(username=username, password=answers["password"])
 
                 except paramiko.ssh_exception.AuthenticationException:
                     print("SSH authentication failed. Not copying files. Continuing")
@@ -261,8 +257,9 @@ def _place_client(src, dest, client, _id):
                     connected = True
 
             else:
-                key = _utils.get_sshkey(expanduser(
-                    answers['ssh_key']), answers['ssh_key_pass'])
+                key = _utils.get_sshkey(
+                    expanduser(answers["ssh_key"]), answers["ssh_key_pass"]
+                )
 
                 if not key:
                     print("\nInvalid ssh key\n")
@@ -279,7 +276,7 @@ def _place_client(src, dest, client, _id):
         sftp = paramiko.SFTPClient.from_transport(transport)
 
         # Make nonexistent paths on remote computer
-        _utils.make_missing_dirs(sftp, dirname, answers['remote_os'])
+        _utils.make_missing_dirs(sftp, dirname, answers["remote_os"])
 
         os.chdir(path_split(client_dir)[0])
         parent = path_split(client_dir)[1]
@@ -290,7 +287,7 @@ def _place_client(src, dest, client, _id):
 
             make_dir = join(dirname, local_subdir)
 
-            if answers['remote_os'] == 'Windows':
+            if answers["remote_os"] == "Windows":
                 make_dir = _utils.nix_to_win_path(make_dir)
             else:
                 make_dir = _utils.win_to_other_path(make_dir)
@@ -301,7 +298,7 @@ def _place_client(src, dest, client, _id):
                 src_file = join(local_subdir, i)
                 dest_file = join(dirname, src_file)
 
-                if answers['remote_os'] == 'Windows':
+                if answers["remote_os"] == "Windows":
                     dest_file = _utils.nix_to_win_path(dest_file)
                 else:
                     dest_file = _utils.win_to_other_path(dest_file)
@@ -315,27 +312,24 @@ def _place_client(src, dest, client, _id):
         sftp.close()
         transport.close()
 
-        return answers['remote_os']
+        return answers["remote_os"]
 
     else:
-        if os.name == 'nt':
-            os_name = 'Windows'
+        if os.name == "nt":
+            os_name = "Windows"
         else:
-            os_name = 'Other'
+            os_name = "Other"
 
-        if client == 'POSIX':
-            dest = join(
-                dest, "NodePingPUSHClient")
+        if client == "POSIX":
+            dest = join(dest, "NodePingPUSHClient")
         else:
-            dest = join(
-                dest, "NodePing{0}PUSH".format(client))
+            dest = join(dest, "NodePing{0}PUSH".format(client))
 
         try:
             os.makedirs(dest)
         except PermissionError:
             print("You do not have permissions to create this directory")
-            print(
-                "You can find the configured client in {0}\n".format(client_dir))
+            print("You can find the configured client in {0}\n".format(client_dir))
             input("Press enter to continue: ")
 
             return os_name
@@ -344,8 +338,7 @@ def _place_client(src, dest, client, _id):
             copy_tree(client_dir, dest)
         except distutils.errors.DistutilsFileError:
             print("You do not have permissions to copy files to this directory")
-            print(
-                "You can find the configured client in {0}\n".format(client_dir))
+            print("You can find the configured client in {0}\n".format(client_dir))
             input("Press enter to continue: ")
 
         return os_name
@@ -355,7 +348,7 @@ def _unzip(zip_archive, dest_dir):
     """ Unzips a .zip archive
     """
 
-    with zipfile.ZipFile(zip_archive, 'r') as master_zip:
+    with zipfile.ZipFile(zip_archive, "r") as master_zip:
         master_zip.extractall(dest_dir)
 
 
@@ -363,55 +356,53 @@ def insert_checktoken(checktoken, _id, unarchived, save_path, client):
     """ Places the checktoken in the client's config file
     """
 
-    if client == 'POSIX':
-        configfile = join(
-            unarchived, 'POSIX/NodePingPUSHClient/NodePingPUSH.sh')
+    if client == "POSIX":
+        configfile = join(unarchived, "POSIX/NodePingPUSHClient/NodePingPUSH.sh")
 
-        with open(configfile, 'r', newline='\n') as f:
+        with open(configfile, "r", newline="\n") as f:
             filedata = f.read()
 
-        moduleconfig = '/full/path/to/moduleconfig'
-        logfile = '/full/path/to/logfile/NodePingPUSH.log'
+        moduleconfig = "/full/path/to/moduleconfig"
+        logfile = "/full/path/to/logfile/NodePingPUSH.log"
 
         # If server is remote, removes the full remote path to just full path
         try:
             full_path = "{0}/NodePingPUSHClient".format(
-                save_path.split('@')[1].split(':')[1])
+                save_path.split("@")[1].split(":")[1]
+            )
         except IndexError:
             full_path = "{0}/NodePingPUSHClient".format(save_path)
 
         # Insert CheckID and checktoken
-        filedata = filedata.replace('CHECK_ID_HERE', _id)
-        filedata = filedata.replace('CHECK_TOKEN_HERE', checktoken)
-        filedata = filedata.replace(
-            moduleconfig, "{0}/moduleconfig".format(full_path))
-        filedata = filedata.replace(
-            logfile, "{0}/NodePingPUSH.log".format(full_path))
+        filedata = filedata.replace("CHECK_ID_HERE", _id)
+        filedata = filedata.replace("CHECK_TOKEN_HERE", checktoken)
+        filedata = filedata.replace(moduleconfig, "{0}/moduleconfig".format(full_path))
+        filedata = filedata.replace(logfile, "{0}/NodePingPUSH.log".format(full_path))
 
-        with open(configfile, 'w', newline='\n') as f:
+        with open(configfile, "w", newline="\n") as f:
             f.write(filedata)
-    elif client == 'Python' or client == 'Python3':
-        configfile = join(
-            unarchived, '{0}/NodePing{0}PUSH/config.ini'.format(client))
+    elif client == "Python" or client == "Python3":
+        configfile = join(unarchived, "{0}/NodePing{0}PUSH/config.ini".format(client))
 
         config = configparser.ConfigParser()
         config.read(configfile)
 
-        config.set('server', 'id', value=_id)
-        config.set('server', 'checktoken', value=checktoken)
+        config.set("server", "id", value=_id)
+        config.set("server", "checktoken", value=checktoken)
 
         _utils.write_config(config, configfile)
-    elif client == 'PowerShell':
+    elif client == "PowerShell":
         configfile = join(
-            unarchived, 'PowerShell/NodePingPowerShellPUSH/NodePingPUSH.ps1')
+            unarchived, "PowerShell/NodePingPowerShellPUSH/NodePingPUSH.ps1"
+        )
 
-        with open(configfile, 'r') as f:
+        with open(configfile, "r") as f:
             filedata = f.read()
 
-        filedata = filedata.replace('Your Check ID here', _id)
-        filedata = filedata.replace('Your Check Token here', checktoken)
+        filedata = filedata.replace("Your Check ID here", _id)
+        filedata = filedata.replace("Your Check Token here", checktoken)
 
-        with open(configfile, 'w') as f:
+        with open(configfile, "w") as f:
             f.write(filedata)
 
 
@@ -419,14 +410,15 @@ def client_set_executable(path, client):
     """
     """
 
-    if client == 'Python' or client == 'Python3':
+    if client == "Python" or client == "Python3":
         # Make Python script executable
         py_script = join(
-            path, "{0}/NodePing{0}PUSH/NodePingPythonPUSH.py".format(client))
+            path, "{0}/NodePing{0}PUSH/NodePingPythonPUSH.py".format(client)
+        )
 
         st = os.stat(py_script)
         os.chmod(py_script, st.st_mode | stat.S_IEXEC)
-    elif client == 'POSIX':
+    elif client == "POSIX":
         # Make POSIX script executable
         sh_script = join(path, "POSIX/NodePingPUSHClient/NodePingPUSH.sh")
 
@@ -445,18 +437,18 @@ def main(metrics, client_zip, client):
     completed_checks = []
 
     # Assigns the token to variables and removes them from the dictionary
-    checktoken = metrics['checktoken']
-    check_id = metrics['check_id']
-    del metrics['checktoken']
-    del metrics['check_id']
+    checktoken = metrics["checktoken"]
+    check_id = metrics["check_id"]
+    del metrics["checktoken"]
+    del metrics["check_id"]
 
     final_destination = _get_save_directory(check_id)
 
     # Assuming the zip file from GitHub is in the same dir as app.py
-    if sys.platform == 'windows':
-        archive_dir = client_zip.strip(client_zip.split('\\')[-1])
+    if sys.platform == "windows":
+        archive_dir = client_zip.strip(client_zip.split("\\")[-1])
     else:
-        archive_dir = client_zip.strip(client_zip.split('/')[-1])
+        archive_dir = client_zip.strip(client_zip.split("/")[-1])
 
     # The name and path of the GH zip file when unarchived
     unarchived = join(archive_dir, CLIENT_UNZIPPED_DIR)
@@ -469,81 +461,100 @@ def main(metrics, client_zip, client):
     _empty_config_file(unarchived, client)
 
     for key, value in metrics.items():
-        name = value['name'].split('.')[0]
+        name = value["name"].split(".")[0]
 
         if name in completed_checks:
             continue
 
-        if 'checksum' in name:
-            name = 'checksum'
-            keys = [key for key in metrics if 'checksum' in key]
+        if "checksum" in name:
+            name = "checksum"
+            keys = [key for key in metrics if "checksum" in key]
 
             import modules.metrics.checksum as checksum
+
             checksum.configure_checksum(keys, metrics, unarchived, client)
 
-        elif 'checkpid' in name:
-            name = 'checkpid'
-            keys = [key for key in metrics if 'checkpid' in key]
+        elif "checkpid" in name:
+            name = "checkpid"
+            keys = [key for key in metrics if "checkpid" in key]
 
             import modules.metrics.checkpid as checkpid
+
             checkpid.configure_checkpid(keys, metrics, unarchived, client)
 
-        elif 'fileage' in name:
-            name = 'fileage'
-            keys = [key for key in metrics if 'fileage' in key]
+        elif "fileage" in name:
+            name = "fileage"
+            keys = [key for key in metrics if "fileage" in key]
 
             import modules.metrics.fileage as fileage
+
             fileage.configure_fileage(keys, metrics, unarchived, client)
 
-        elif 'pingstatus' in name:
-            name = 'pingstatus'
-            keys = [key for key in metrics if 'pingstatus' in key]
+        elif "pingstatus" in name:
+            name = "pingstatus"
+            keys = [key for key in metrics if "pingstatus" in key]
 
             import modules.metrics.pingstatus as pingstatus
+
             pingstatus.configure_pingstatus(keys, metrics, unarchived, client)
 
-        elif 'httpcheck' in name:
-            name = 'httpcheck'
+        elif "httpcheck" in name:
+            name = "httpcheck"
             keys = [key for key in metrics if name in key]
 
             # configure_httpcheck(keys, metrics, unarchived, client)
             import modules.metrics.httpcheck as httpcheck
+
             httpcheck.configure_httpcheck(unarchived, client)
 
-        elif name == 'mysqlstat':
+        elif name == "mysqlstat":
             import modules.metrics.sqlstat as sqlstat
+
             sqlstat.configure_mysqlstat(unarchived, client)
 
-        elif name == 'pgsqlstat':
+        elif name == "pgsqlstat":
             import modules.metrics.sqlstat as sqlstat
+
             sqlstat.configure_pgsqlstat(unarchived, client)
 
-        elif name == 'redismaster':
+        elif name == "redismaster":
             import modules.metrics.redismaster as redismaster
+
             redismaster.configure_redismaster(unarchived, client)
 
-        elif name == 'ip_addrs':
+        elif name == "ip_addrs":
             import modules.metrics.ip_addrs as ip_addrs
+
             ip_addrs.configure_ip_addrs(unarchived, client)
 
-        elif name == 'dnslookup':
+        elif name == "dnslookup":
             import modules.metrics.dnslookup as dnslookup
+
             dnslookup.configure_dnslookup(unarchived, client)
 
-        elif name == 'mongodbstat':
+        elif name == "mongodbstat":
             import modules.metrics.mongodbstat as mongodbstat
+
             mongodbstat.configure_mongodbstat(unarchived, client)
 
-        elif name == 'smartctl':
-            keys = [key for key in metrics if 'smartctl' in key]
+        elif name == "net_utilization":
+            keys = [key for key in metrics if "net_utilization" in key]
+            import modules.metrics.net_utilization as net_utilization
+
+            net_utilization.configure_net_utilization(keys, metrics, unarchived, client)
+
+        elif name == "smartctl":
+            keys = [key for key in metrics if "smartctl" in key]
 
             import modules.metrics.smartctl as smartctl
+
             smartctl.configure_smartctl(keys, metrics, unarchived, client)
 
-        elif name == 'zpool':
-            keys = [key for key in metrics if 'zpool' in key]
+        elif name == "zpool":
+            keys = [key for key in metrics if "zpool" in key]
 
             import modules.metrics.zpool as zpool
+
             zpool.configure_zpool(keys, metrics, unarchived, client)
 
         _add_metric(unarchived, final_destination, name, client)
@@ -551,8 +562,7 @@ def main(metrics, client_zip, client):
         completed_checks.append(name)
 
     # Inserts check token into the proper file
-    insert_checktoken(checktoken, check_id, unarchived,
-                      final_destination, client)
+    insert_checktoken(checktoken, check_id, unarchived, final_destination, client)
 
     # Sets client script to executable
     client_set_executable(unarchived, client)
@@ -560,9 +570,9 @@ def main(metrics, client_zip, client):
     # Gets OS that client was copied to
     user_os = _place_client(unarchived, final_destination, client, check_id)
 
-    if '@' in final_destination:
-        dest = final_destination.split('@')[1].split(':')[1]
-        user = final_destination.split('@')[0]
-        return {'user': user, 'dest': dest, 'os': user_os}
+    if "@" in final_destination:
+        dest = final_destination.split("@")[1].split(":")[1]
+        user = final_destination.split("@")[0]
+        return {"user": user, "dest": dest, "os": user_os}
 
-    return {'user': '', 'dest': final_destination, 'os': user_os}
+    return {"user": "", "dest": final_destination, "os": user_os}